requirements.txt
python-telegram-bot
edge-tts
python-dotenv